requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
flask>=3.0.0
flask-cors>=5.0.0
